refactor(vscode): reuse string truncation helper
